feat(phase10): consolidate API, update imports, zero TypeScript errors
T034: Create shared/ai/index.ts — main barrel file re-exporting all
  modules (catalog, storage, paths, download, validation, manager,
  runtime, types, constants, errors)

T035: Remove legacy barrel files (model-manager.ts, model-catalog.ts,
  local-ai-runtime.ts) — now imports point to actual modules

T036: Update all external imports in features/ to use new paths:
  - use-chat-vm.ts → @/shared/ai/runtime/local-ai-runtime.service
  - use-models-vm.ts → @/shared/ai/catalog, @/shared/ai/manager/...
  - models-screen.tsx → @/shared/ai/catalog
  - model-catalog.tsx → @/shared/ai/catalog

T037: Zero TypeScript errors (npx tsc --noEmit passes cleanly)

Refs: #116
